add root provider
